+++ internship/models/internship_student_affectation_stat.py
@@ -23,7 +23,14 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+from django.contrib import admin
 from django.db import models
+
+
+class InternshipStudentAffectationStatAdmin(admin.ModelAdmin):
+    list_display = ('student', 'organization', 'speciality', 'period', 'choice', 'cost', 'consecutive_month', 'type_of_internship')
+    fieldsets = ((None, {'fields': ('student', 'organization', 'speciality', 'period', 'choice', 'cost', 'consecutive_month', 'type_of_internship')}),)
+    raw_id_fields = ('student', 'organization', 'speciality', 'period')
 
 
 class InternshipStudentAffectationStat(models.Model):
